requirements.txt
requests>=2.28.0
PyPDF2>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
flask>=3.0.0
flask-cors>=4.0.0
python-docx>=1.0.0
